+++ src/game/system/parsing/PlayerCommand.java
@@ -6,7 +6,7 @@ import util.FuncUtils;
 
 /**
  * Represents a command that the player has issued to the game in a form of a
- * {@link String}. A player command is composed of a series of {@link Action}s
+ * {@link String}. A player command is composed of a series of {@link PlayerAction}s
  * that are sequentially ordered.
  *
  * @author Evan Quan
@@ -20,9 +20,9 @@ public class PlayerCommand {
     private String string;
 
     /**
-     * Series of actions in the order they were issued.
+     * Series of playerActions in the order they were issued.
      */
-    ArrayList<Action> actions;
+    ArrayList<PlayerAction> playerActions;
 
     /**
      * Creates a Player Command instance.
@@ -31,20 +31,20 @@ public class PlayerCommand {
      */
     public PlayerCommand(String string) {
         this.string = string;
-        actions = new ArrayList<>();
+        playerActions = new ArrayList<>();
     }
 
     /**
-     * Add a single action to this commands list of actions.
+     * Add a single playerAction to this commands list of playerActions.
      * 
-     * @param action
+     * @param playerAction
      */
-    public void addAction(Action action) {
-        this.actions.add(action);
+    public void addAction(PlayerAction playerAction) {
+        this.playerActions.add(playerAction);
     }
 
-    public boolean contains(Action action) {
-        return this.actions.contains(action);
+    public boolean contains(PlayerAction playerAction) {
+        return this.playerActions.contains(playerAction);
     }
 
     @Override
@@ -55,8 +55,8 @@ public class PlayerCommand {
         return false;
     }
 
-    public ArrayList<Action> getActions() {
-        return this.actions;
+    public ArrayList<PlayerAction> getPlayerActions() {
+        return this.playerActions;
     }
 
     public String getString() {
@@ -64,29 +64,29 @@ public class PlayerCommand {
     }
 
     public boolean hasActions() {
-        return this.actions != null && !this.actions.isEmpty();
+        return this.playerActions != null && !this.playerActions.isEmpty();
     }
 
     public boolean hasSameActions(PlayerCommand other) {
-        return FuncUtils.nullablesEqual(this.actions, other.getActions());
+        return FuncUtils.nullablesEqual(this.playerActions, other.getPlayerActions());
     }
 
     public boolean hasSameString(PlayerCommand other) {
         return FuncUtils.nullablesEqual(this.string, other.getString());
     }
 
-    public void setActions(ArrayList<Action> actions) {
-        this.actions = actions;
+    public void setPlayerActions(ArrayList<PlayerAction> playerActions) {
+        this.playerActions = playerActions;
     }
 
     @Override
     public String toString() {
         StringBuilder actions = new StringBuilder();
         if (hasActions()) {
-            for (Action action : this.actions) {
-                actions.append("\t" + action + "\n");
+            for (PlayerAction playerAction : this.playerActions) {
+                actions.append("\t" + playerAction + "\n");
             }
         }
-        return "[string: " + string + "\nactions: " + actions + "]";
+        return "[string: " + string + "\nplayerActions: " + actions + "]";
     }
 }