+++ src/game/system/parsing/VerbAgnosticPlayerCommandParser.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 import game.system.parsing.words.ObjectPhrase;
+import game.system.parsing.words.VerbPhrase;
 import game.system.parsing.words.Word;
 import util.ArrayUtils;
 
@@ -27,7 +28,7 @@ import util.ArrayUtils;
  * <p>
  * <b>Due to not having a verb dictionary, this cannot do the following:</b>
  * <p>
- * Multiple action commands, such as:<br>
+ * Multiple playerAction commands, such as:<br>
  * Multiverb commands: (look up, eat pie, go west)<br>
  * Verbsharing commands: (eat pie, potato, cake)<br>
  * Object pronouns (this may not be implemented here): (take pie, eat it)<br>
@@ -39,9 +40,9 @@ import util.ArrayUtils;
  * - A verb dictionary<br>
  * - lexicalAnalysis() needs to recognize commas at the end of words as their
  * own tokens<br>
- * - syntacticalAnalys() needs to separate actions by separators<br>
- * - incomplete actions need to be able to "fill in the gaps" from context of
- * previously parsed actions in the same command<br>
+ * - syntacticalAnalys() needs to separate playerActions by separators<br>
+ * - incomplete playerActions need to be able to "fill in the gaps" from context of
+ * previously parsed playerActions in the same command<br>
  *
  * @author Evan Quan
  *
@@ -218,10 +219,10 @@ public abstract class VerbAgnosticPlayerCommandParser {
             // This happens when the player input an empty string
             return;
         }
-        // TODO when multi-action commands are implemented, make this part a loop for
+        // TODO when multi-playerAction commands are implemented, make this part a loop for
         // every separator section
 
-        Action action = new Action();
+        PlayerAction playerAction = new PlayerAction();
 
         String first = tokens.get(0);
         if (!Word.isDeterminer(first) && !Word.isPreposition(first)) {
@@ -230,7 +231,7 @@ public abstract class VerbAgnosticPlayerCommandParser {
             // end of the verb phrase and the start of the proceeding indirect/direct object
             // phrase without a dictionary of all possible verbs.
             tokens.remove(0);
-            action.setVerbPhrase(first);
+            playerAction.setVerbPhrase(new VerbPhrase(first));
         }
         // 1. Scan for a preposition. If one is found, remove it. Parse the input
         // preceding the preposition as a direct object phrase. Parse the input
@@ -246,7 +247,7 @@ public abstract class VerbAgnosticPlayerCommandParser {
         for (i = 0; i < tokens.size(); i++) {
             String token = tokens.get(i);
             if (Word.isPreposition(token)) {
-                action.setPreposition(token);
+                playerAction.setPreposition(token);
                 break;
             } else {
                 directTokens.add(token);
@@ -259,11 +260,11 @@ public abstract class VerbAgnosticPlayerCommandParser {
         }
 
         // Create the object phrases from the token lists
-        action.setDirectObjectPhrase(getObjectPhrase(directTokens));
-        action.setIndirectObjectPhrase(getObjectPhrase(indirectTokens));
+        playerAction.setDirectObjectPhrase(getObjectPhrase(directTokens));
+        playerAction.setIndirectObjectPhrase(getObjectPhrase(indirectTokens));
 
-        // Add complete action to player command
-        playerCommand.addAction(action);
+        // Add complete playerAction to player command
+        playerCommand.addAction(playerAction);
     }
 
     // /**
@@ -271,7 +272,7 @@ public abstract class VerbAgnosticPlayerCommandParser {
     // * <p>
     // * TODO Creates a player command from a list of words. Depending on the
     // relation
-    // * between words, the action {@link Verb} and object {@link Noun} are
+    // * between words, the playerAction {@link Verb} and object {@link Noun} are
     // * determined.
     // *
     // * @param input
@@ -284,9 +285,9 @@ public abstract class VerbAgnosticPlayerCommandParser {
     // int actionIndex = 0;
     // int objectIndex = 0;
     // // 1. The first word of the command should either be a verb, or a shortcut
-    // // represents some action
+    // // represents some playerAction
     //
-    // // return new PlayerCommand(command, action, object);
+    // // return new PlayerCommand(command, playerAction, object);
     // return null;
     // }
 }